ai-service/requirements.txt
flask==3.1.0
flask-cors==5.0.0
pdfplumber==0.11.4
pymupdf==1.24.14
pytesseract==0.3.13
Pillow==11.0.0
pdf2image==1.17.0
scikit-learn==1.6.0
numpy==2.2.1
joblib==1.4.2
requests==2.32.3
python-dotenv==1.0.1
